# cmd/workload/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Gin swagger base",
        "title": "Gin swagger",
        "contact": {
            "name": "laik author",
            "url": "github.com/yametech",
            "email": "laik.lj@me.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
        "/workload/v1/:ns/pod/:name/get": {
            "get": {
                "description": "workload service for get a pod detail",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pod"
                ],
                "summary": "workload pod get",
                "parameters": [
                    {
                        "type": "string",
                        "description": "namespace",
                        "name": "namespace",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "name",
                        "name": "name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"msg\": \"Success\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"msg\": \"Failed\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/workload/v1/:ns/pod/list": {
            "get": {
                "description": "workload service for list pod",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pod"
                ],
                "summary": "workload pod list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "namespace",
                        "name": "namespace",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"msg\": \"Success\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"msg\": \"Failed\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}